add fs

+++ 10-file-system/04-my-rm.js
@@ -3,5 +3,10 @@
 const fs = require('fs'),
       file = process.argv[2];
 
-fs.unlinkSync(file);
+if(fs.existsSync(file)){
+    if(fs.statSync(file).isFile()) fs.unlinkSync(file);
+}else{
+    console.error('not exist');
+    process.exit(1);
+}
